new example format
