+++ solutions/mint-nft/.eslintrc.json
