+++ solutions/mint-nft/.gitignore
