+++ solutions/mint-nft/helpers/address.helpers.ts
