+++ solutions/mint-nft/helpers/constant.helpers.ts
