+++ solutions/mint-nft/lib/fetcher.ts
