+++ solutions/mint-nft/next-env.d.ts
