+++ solutions/mint-nft/package.json
@@ -1,6 +1,8 @@
 {
-  "name": "nft-mint",
-  "private": true,
+  "name": "vercel-example--mint-nft",
+  "version": "1.0.0",
+  "repository": "https://github.com/vercel/examples.git",
+  "license": "MIT",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -21,18 +23,16 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-dropzone": "^11.5.3",
-    "react-intersection-observer": "^8.33.1",
     "react-lottie-player": "^1.4.1",
-    "react-moralis": "^1.2.1",
-    "swr": "^1.2.1",
+    "react-moralis": "^1.3.1",
     "wagmi": "^0.2.8"
   },
   "devDependencies": {
     "@types/node": "17.0.14",
-    "@types/react": "17.0.39",
+    "@types/react": "17.0.38",
     "autoprefixer": "^10.4.2",
-    "eslint": "8.8.0",
-    "eslint-config-next": "12.0.10",
+    "eslint": "^8.9.0",
+    "eslint-config-next": "^12.0.10",
     "postcss": "^8.4.6",
     "tailwindcss": "^3.0.18",
     "typescript": "4.5.5"
