+++ solutions/mint-nft/tailwind.config.js
