+++ solutions/mint-nft/tsconfig.json
